+++ training/create_model.py
@@ -5,7 +5,7 @@ import pandas as pd
 from lightgbm import early_stopping, log_evaluation
 from mord import LogisticAT, LogisticIT
 from orf import OrderedForest
-from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
+from sklearn.ensemble import RandomForestRegressor
 from sklearn.linear_model import (
     HuberRegressor,
     LassoCV,
@@ -21,8 +21,8 @@ from statsmodels.miscmodels.ordinal_model import OrderedModel
 
 from training.constants import RANDOM_STATE
 from training.models.gpor import GPOR
+from training.models.ordered_models import LinearOrdinalModel
 from training.models.simple_ordinal_classification import SimpleOrdinalClassification
-from training.ordered_models import LinearOrdinalModel
 from training.score_functions import orf_mean_absolute_error
 
 
@@ -187,7 +187,6 @@ def create_model(classifier_name: str):
                 return_train_score=True,
                 n_jobs=-1,
             )
-        
         case "linear_ordinal_model_probit":
             model = create_linear_ordinal_model("probit")
         case "linear_ordinal_model_logit":
